refactor(processors): split user, files, and repository at their seams
Pure moves, no behavior change — each extracted file is a nameable system
that had accreted into the processor:

user.go (1059 → 376): the shadow-utils backend (every *Linux function,
setLinuxPassword, isCryptHash) moves to user_shadowutils.go and the Open
Directory backend (every *Darwin function, the mac* helpers) to
user_opendirectory.go. Backend names, not _linux/_darwin suffixes — those
carry implicit GOOS build constraints and this code dispatches at runtime
on the detected OS. user_test.go splits the same way (845 shadow-utils,
317 Open Directory, 131 shared fixtures + orchestration); all 44 tests
unchanged and passing.

files.go (874 → 519): directories (processDirectory + the eight directory
verbs + attribute application) move to directories.go, templates to
templates.go. The per-file action verbs stay — they are the file
processor, not a separable system.

repository.go (733 → 310): the two embedded subsystems move out — step
machinery (renderActionStep, repositoryStepData, predicates, conditions)
to repository_steps.go, and the path-containment layer (repositoryWritePath,
boundaries, removal resolution, withinAny) to repository_paths.go, putting
the security-sensitive code in one reviewable file.

ssh_key.go stays: one coherent flow that happens to be long.
