+++ internal/processors/files.go
@@ -255,184 +255,6 @@ func processFile(file types.File, blueprintDir string, osInfo *types.OSInfo) err
 	}
 }
 
-func processTemplates(templates []types.File, blueprintDir string, osInfo *types.OSInfo, initConfig *types.InitConfig, track *progress) error {
-	log.Info("Starting to process templates")
-
-	total := 0
-	for _, tmpl := range templates {
-		if len(tmpl.Names) > 0 {
-			total += len(tmpl.Names)
-		} else if tmpl.Name != "" {
-			total++
-		}
-	}
-	track.expect("", total)
-
-	run := func(tmpl types.File) {
-		started := time.Now()
-		switch err := processTemplate(tmpl, blueprintDir, osInfo, initConfig); {
-		case err != nil:
-			recordFailure("templates", tmpl.Name, err)
-			track.item("", tmpl.Name, "template", types.StatusFailed, err.Error(), time.Since(started))
-		case system.IsDryRun():
-			track.item("", tmpl.Name, "template", types.StatusPlanned, "dry-run", 0)
-		case tmpl.Source == "" || tmpl.Target == "":
-			// processTemplate warns and returns nil for these; mirror that as a
-			// skip rather than a success.
-			track.item("", tmpl.Name, "template", types.StatusSkipped, "missing required fields", 0)
-		default:
-			track.itemIdentity("", tmpl.Name, "template", types.StatusOK, "", time.Since(started), fileJournalIdentity(tmpl, blueprintDir))
-		}
-	}
-
-	for i, tmpl := range templates {
-		log.Debugf("Processing template %d: %+v", i, tmpl)
-		if tmpl.Name == "" && len(tmpl.Names) == 0 {
-			log.Warn("Skipping empty template")
-			continue
-		}
-		if len(tmpl.Names) > 0 {
-			log.Debugf("Template has multiple names: %v", tmpl.Names)
-			for _, name := range tmpl.Names {
-				log.Infof("Processing template with name: %s", name)
-				fileWithName := tmpl
-				fileWithName.Name = name
-				run(fileWithName)
-			}
-		} else {
-			log.Infof("Processing single template: %s", tmpl.Name)
-			run(tmpl)
-		}
-	}
-	log.Info("Finished processing all templates")
-	return nil
-}
-
-func processTemplate(template types.File, blueprintDir string, osInfo *types.OSInfo, initConfig *types.InitConfig) error {
-	log.Infof("Processing template: %s", template.Name)
-
-	if template.Name == "" || template.Source == "" || template.Target == "" {
-		log.Warnf("Skipping template with missing required fields: %+v", template)
-		return nil
-	}
-
-	sourcePath := filepath.Join(blueprintDir, template.Source, template.Name)
-	log.Debugf("Full source path: %s", sourcePath)
-
-	content, err := os.ReadFile(sourcePath) // #nosec G304 -- path is operator-supplied blueprint/config input; containment added in PR8
-	if err != nil {
-		log.Errorf("Error reading template file %s: %v", sourcePath, err)
-		return fmt.Errorf("error reading template file %s: %w", sourcePath, err)
-	}
-	log.Debugf("Successfully read template file, content length: %d bytes", len(content))
-
-	log.Debug("Resolving template variables")
-	// Copying the struct still shares the UserDefined map, so writing template
-	// overrides into it mutated initConfig for every later template and
-	// blueprint — a per-template variable permanently clobbered a run-wide one
-	// of the same name. Merge into a fresh map instead.
-	mergedVariables := initConfig.Variables
-	mergedVariables.UserDefined = make(map[string]interface{}, len(initConfig.Variables.UserDefined)+len(template.Variables))
-	for k, v := range initConfig.Variables.UserDefined {
-		mergedVariables.UserDefined[k] = v
-	}
-	for k, v := range template.Variables {
-		mergedVariables.UserDefined[k] = v
-	}
-	resolvedContent, err := helpers.ResolveTemplate(content, mergedVariables)
-	if err != nil {
-		log.Errorf("Error resolving template %s: %v", sourcePath, err)
-		return fmt.Errorf("error resolving template %s: %w", sourcePath, err)
-	}
-	log.Debugf("Successfully resolved template, new content length: %d bytes", len(resolvedContent))
-
-	mode := template.Mode
-	if !mode.IsSet() {
-		mode = types.FileMode(defaultTemplateMode)
-	}
-
-	// Create a File struct from the Template
-	file := types.File{
-		Name:     template.Name,
-		Action:   template.Action,
-		Content:  string(resolvedContent),
-		Target:   template.Target,
-		Owner:    template.Owner,
-		Group:    template.Group,
-		Mode:     mode,
-		Elevated: template.Elevated,
-	}
-
-	// Process the template as a file
-	err = processFile(file, blueprintDir, osInfo)
-	if err != nil {
-		log.Errorf("Error processing template as file %s: %v", template.Name, err)
-		return fmt.Errorf("error processing template as file %s: %w", template.Name, err)
-	}
-
-	log.Infof("Template processed successfully: %s", template.Name)
-	return nil
-}
-
-func processDirectories(directories []types.Directory, blueprintDir string, initConfig *types.InitConfig, track *progress) error {
-	track.expect("", len(directories))
-	for _, dir := range directories {
-		if system.IsDryRun() {
-			log.Infof("[DRY-RUN] Would %s directory: %s (target: %s)", dir.Action, dir.Name, dir.Target)
-			track.item("", dir.Name, dir.Action, types.StatusPlanned, "dry-run", 0)
-			continue
-		}
-		started := time.Now()
-		if err := processDirectory(dir, blueprintDir, initConfig); err != nil {
-			recordFailure("directories", dir.Name, err)
-			track.item("", dir.Name, dir.Action, types.StatusFailed, err.Error(), time.Since(started))
-			continue
-		}
-		track.itemIdentity("", dir.Name, dir.Action, types.StatusOK, "", time.Since(started), map[string]string{"dest": system.ExpandPath(resolveTargetPath(dir.Target, dir.Name))})
-	}
-	return nil
-}
-
-func processDirectory(dir types.Directory, blueprintDir string, initConfig *types.InitConfig) error {
-	switch dir.Action {
-	case "copy":
-		if err := copyDirectory(dir, blueprintDir, initConfig); err != nil {
-			return fmt.Errorf("error copying directory: %w", err)
-		}
-	case "move":
-		if err := moveDirectory(dir, blueprintDir); err != nil {
-			return fmt.Errorf("error moving directory: %w", err)
-		}
-	case "delete":
-		if err := deleteDirectory(dir); err != nil {
-			return fmt.Errorf("error deleting directory: %w", err)
-		}
-	case "create":
-		if err := createDirectory(dir); err != nil {
-			return fmt.Errorf("error creating directory: %w", err)
-		}
-	case "chmod":
-		if err := chmodDirectory(dir); err != nil {
-			return fmt.Errorf("error changing directory permissions: %w", err)
-		}
-	case "chown":
-		if err := chownDirectory(dir); err != nil {
-			return fmt.Errorf("error changing directory owner: %w", err)
-		}
-	case "chgrp":
-		if err := chgrpDirectory(dir); err != nil {
-			return fmt.Errorf("error changing directory group: %w", err)
-		}
-	case "symlink":
-		if err := symlinkDirectory(dir, blueprintDir); err != nil {
-			return fmt.Errorf("error creating symlink: %w", err)
-		}
-	default:
-		return fmt.Errorf("unsupported action for directory: %s", dir.Action)
-	}
-	return nil
-}
-
 func moveFile(source, target string) error {
 	if err := os.MkdirAll(filepath.Dir(target), defaultDirMode); err != nil {
 		return fmt.Errorf("error creating target directory: %w", err)
@@ -617,139 +439,6 @@ func ensureSymlink(source, target string) error {
 	return nil
 }
 
-func copyDirectory(dir types.Directory, blueprintDir string, initConfig *types.InitConfig) error {
-	source := filepath.Join(blueprintDir, dir.Source, dir.Name)
-	target := filepath.Join(system.ExpandPath(dir.Target), dir.Name)
-
-	if err := os.MkdirAll(target, directoryMode(dir)); err != nil { // #nosec G703 -- target path is operator-supplied blueprint/config input; containment added in PR8
-		return fmt.Errorf("error creating target directory: %w", err)
-	}
-
-	if err := system.CopyDirectory(source, target, dir.Elevated, helpers.ResolveInteractive(dir.Interactive, initConfig.Variables.Flags.Interactive)); err != nil {
-		return fmt.Errorf("error copying directory: %w", err)
-	}
-
-	if err := applyDirectoryAttributes(dir); err != nil {
-		return fmt.Errorf("error applying directory attributes: %w", err)
-	}
-
-	log.Infof("Directory copied: %s -> %s", source, target)
-	return nil
-}
-
-func moveDirectory(dir types.Directory, blueprintDir string) error {
-	source := filepath.Join(blueprintDir, dir.Source, dir.Name)
-	target := filepath.Join(system.ExpandPath(dir.Target), dir.Name)
-
-	targetDir := filepath.Dir(target)
-	if err := os.MkdirAll(targetDir, defaultDirMode); err != nil {
-		return fmt.Errorf("error creating target directory: %w", err)
-	}
-
-	if err := os.Rename(source, target); err != nil {
-		if os.IsNotExist(err) {
-			if _, statErr := os.Lstat(target); statErr == nil {
-				log.Infof("Directory already moved: %s", target)
-				return nil
-			}
-		}
-		return fmt.Errorf("error moving directory: %w", err)
-	}
-
-	log.Infof("Directory moved: %s -> %s", source, target)
-	return nil
-}
-
-func deleteDirectory(dir types.Directory) error {
-	target := filepath.Join(system.ExpandPath(dir.Target), dir.Name)
-
-	if err := os.RemoveAll(target); err != nil {
-		return fmt.Errorf("error deleting directory: %w", err)
-	}
-
-	log.Infof("Directory deleted: %s", target)
-	return nil
-}
-
-func createDirectory(dir types.Directory) error {
-	target := filepath.Join(system.ExpandPath(dir.Target), dir.Name)
-
-	if err := os.MkdirAll(target, directoryMode(dir)); err != nil { // #nosec G703 -- target path is operator-supplied blueprint/config input; containment added in PR8
-		return fmt.Errorf("error creating directory: %w", err)
-	}
-
-	if err := applyDirectoryAttributes(dir); err != nil {
-		return fmt.Errorf("error applying directory attributes: %w", err)
-	}
-
-	log.Infof("Directory created: %s", target)
-	return nil
-}
-
-func chmodDirectory(dir types.Directory) error {
-	target := filepath.Join(system.ExpandPath(dir.Target), dir.Name)
-
-	if !dir.Mode.IsSet() {
-		return fmt.Errorf("no mode declared for chmod of %s: add mode: \"0755\" to the directory entry", target)
-	}
-
-	if err := os.Chmod(target, dir.Mode.OSMode()); err != nil {
-		return fmt.Errorf("error changing directory permissions: %w", err)
-	}
-
-	log.Infof("Directory permissions changed: %s (mode: %s)", target, dir.Mode)
-	return nil
-}
-
-func chownDirectory(dir types.Directory) error {
-	target := filepath.Join(system.ExpandPath(dir.Target), dir.Name)
-
-	if dir.Owner != "" {
-		uid, err := system.LookupUID(dir.Owner)
-		if err != nil {
-			return fmt.Errorf("error looking up owner UID: %w", err)
-		}
-		if err := os.Chown(target, uid, -1); err != nil {
-			return fmt.Errorf("error changing directory owner: %w", err)
-		}
-	}
-
-	if dir.Group != "" {
-		gid, err := system.LookupGID(dir.Group)
-		if err != nil {
-			return fmt.Errorf("error looking up group GID: %w", err)
-		}
-		if err := os.Chown(target, -1, gid); err != nil {
-			return fmt.Errorf("error changing directory group: %w", err)
-		}
-	}
-
-	log.Infof("Directory owner/group changed: %s (owner: %s, group: %s)", target, dir.Owner, dir.Group)
-	return nil
-}
-
-// chgrpDirectory is the group half of chownDirectory: same lookup, same chown
-// call, owner left untouched.
-func chgrpDirectory(dir types.Directory) error {
-	groupOnly := dir
-	groupOnly.Owner = ""
-	return chownDirectory(groupOnly)
-}
-
-func symlinkDirectory(dir types.Directory, blueprintDir string) error {
-	source := filepath.Join(blueprintDir, dir.Source, dir.Name)
-	target := system.ExpandPath(dir.Target)
-
-	return ensureSymlink(source, target)
-}
-
-func directoryMode(dir types.Directory) os.FileMode {
-	if dir.Mode.IsSet() {
-		return dir.Mode.OSMode()
-	}
-	return defaultDirMode
-}
-
 func applyFileAttributes(targetPath string, file types.File) error {
 	if file.Mode.IsSet() {
 		if err := os.Chmod(targetPath, file.Mode.OSMode()); err != nil { // #nosec G703 -- target path is operator-supplied blueprint/config input; containment added in PR8
@@ -760,24 +449,6 @@ func applyFileAttributes(targetPath string, file types.File) error {
 	if file.Owner != "" || file.Group != "" {
 		if err := chownFile(file, targetPath); err != nil {
 			return fmt.Errorf("error changing file owner/group: %v", err)
-		}
-	}
-
-	return nil
-}
-
-func applyDirectoryAttributes(dir types.Directory) error {
-	target := filepath.Join(system.ExpandPath(dir.Target), dir.Name)
-
-	if dir.Mode.IsSet() {
-		if err := os.Chmod(target, dir.Mode.OSMode()); err != nil {
-			return fmt.Errorf("error changing directory permissions: %w", err)
-		}
-	}
-
-	if dir.Owner != "" || dir.Group != "" {
-		if err := chownDirectory(dir); err != nil {
-			return fmt.Errorf("error changing directory owner/group: %w", err)
 		}
 	}
 
@@ -844,31 +515,5 @@ func processFileImports(items []types.File, blueprintDir string, format string, 
 				return nil, err
 			}
 			return d.Files, nil
-		}, format)
-}
-
-// processTemplateImports takes the imported file's `templates:` list. Taking its
-// `files:` list instead loses every imported template without an error.
-func processTemplateImports(items []types.File, blueprintDir string, format string, treeVersion int) ([]types.File, error) {
-	return helpers.ResolveImports(items, blueprintDir,
-		func(item types.File) string { return item.Import },
-		func(data []byte, fileFormat string) ([]types.File, error) {
-			var d types.FileData
-			if err := helpers.DecodeBlueprintInto(data, fileFormat, types.BlueprintTypeFiles, treeVersion, &d); err != nil {
-				return nil, err
-			}
-			return d.Templates, nil
-		}, format)
-}
-
-func processDirectoryImports(items []types.Directory, blueprintDir string, format string, treeVersion int) ([]types.Directory, error) {
-	return helpers.ResolveImports(items, blueprintDir,
-		func(item types.Directory) string { return item.Import },
-		func(data []byte, fileFormat string) ([]types.Directory, error) {
-			var d types.FileData
-			if err := helpers.DecodeBlueprintInto(data, fileFormat, types.BlueprintTypeFiles, treeVersion, &d); err != nil {
-				return nil, err
-			}
-			return d.Directories, nil
 		}, format)
 }
