+++ internal/processors/user.go
@@ -4,8 +4,6 @@ import (
 	"fmt"
 	"regexp"
 	"runtime"
-	"strconv"
-	"strings"
 	"time"
 
 	"charm.land/log/v2"
@@ -236,19 +234,6 @@ func userExists(name string, initConfig *types.InitConfig, assumeDryRun bool) bo
 	return system.RunCommand(cmd, initConfig.Variables.Flags.Debug) == nil
 }
 
-// macUserInGroup uses dseditgroup's own membership query. dseditgroup -o edit -a
-// fails when the user is already a member, so the add has to be conditional.
-func macUserInGroup(user, group string, initConfig *types.InitConfig) bool {
-	if system.IsDryRun() {
-		return false
-	}
-	cmd := types.Command{
-		Exec: "dseditgroup",
-		Args: []string{"-o", "checkmember", "-m", user, group},
-	}
-	return system.RunCommand(cmd, initConfig.Variables.Flags.Debug) == nil
-}
-
 // ---------------------------------------------------------------------------
 // groups
 // ---------------------------------------------------------------------------
@@ -267,66 +252,6 @@ func createGroup(group types.Group, initConfig *types.InitConfig) error {
 	return nil
 }
 
-func createGroupLinux(group types.Group, initConfig *types.InitConfig) error {
-	if groupExists(group.Name, initConfig, false) {
-		log.Infof("Group %s already exists, converging attributes instead of creating", group.Name)
-		return modifyGroup(types.Group{Name: group.Name, GID: group.GID}, initConfig)
-	}
-
-	args := []string{}
-	if group.GID != "" {
-		args = append(args, "--gid", group.GID)
-	}
-	if group.System {
-		args = append(args, "--system")
-	}
-	args = append(args, group.Name)
-
-	createGroupCmd := types.Command{
-		Exec:     "groupadd",
-		Args:     args,
-		Elevated: true,
-	}
-	if err := system.RunCommand(createGroupCmd, initConfig.Variables.Flags.Debug); err != nil {
-		return fmt.Errorf("error creating group: %v", err)
-	}
-	return nil
-}
-
-func createGroupDarwin(group types.Group, initConfig *types.InitConfig) error {
-	if group.System {
-		log.Warnf("Group %s: 'system' is ignored on macOS; GIDs below %d are reserved by Apple and are not allocated by rwr", group.Name, macMinID)
-	}
-
-	if groupExists(group.Name, initConfig, false) {
-		log.Infof("Group %s already exists, converging attributes instead of creating", group.Name)
-		return modifyGroup(types.Group{Name: group.Name, GID: group.GID}, initConfig)
-	}
-
-	gid := group.GID
-	if gid == "" {
-		allocated, err := macNextFreeID("/Groups", "PrimaryGroupID", initConfig)
-		if err != nil {
-			return fmt.Errorf("error allocating GID for group %s: %w", group.Name, err)
-		}
-		gid = allocated
-	}
-
-	record := "/Groups/" + group.Name
-	steps := [][]string{
-		{".", "-create", record},
-		{".", "-create", record, "PrimaryGroupID", gid},
-		{".", "-create", record, "RealName", group.Name},
-	}
-	for _, args := range steps {
-		cmd := types.Command{Exec: "dscl", Args: args, Elevated: true}
-		if err := system.RunCommand(cmd, initConfig.Variables.Flags.Debug); err != nil {
-			return fmt.Errorf("error creating group: %v", err)
-		}
-	}
-	return nil
-}
-
 func modifyGroup(group types.Group, initConfig *types.InitConfig) error {
 	switch userGOOS {
 	case "linux":
@@ -337,58 +262,6 @@ func modifyGroup(group types.Group, initConfig *types.InitConfig) error {
 		log.Warnf("Modifying groups is not supported on Windows")
 	default:
 		return fmt.Errorf("unsupported operating system: %s", userGOOS)
-	}
-	return nil
-}
-
-func modifyGroupLinux(group types.Group, initConfig *types.InitConfig) error {
-	if group.NewName == "" && group.GID == "" {
-		log.Debugf("Group %s: nothing to modify", group.Name)
-		return nil
-	}
-
-	modifyGroupCmd := types.Command{
-		Exec:     "groupmod",
-		Args:     []string{group.Name},
-		Elevated: true,
-	}
-	if group.NewName != "" {
-		modifyGroupCmd.Args = append(modifyGroupCmd.Args, "--new-name", group.NewName)
-	}
-	if group.GID != "" {
-		modifyGroupCmd.Args = append(modifyGroupCmd.Args, "--gid", group.GID)
-	}
-
-	if err := system.RunCommand(modifyGroupCmd, initConfig.Variables.Flags.Debug); err != nil {
-		return fmt.Errorf("error modifying group: %v", err)
-	}
-	return nil
-}
-
-func modifyGroupDarwin(group types.Group, initConfig *types.InitConfig) error {
-	record := "/Groups/" + group.Name
-
-	if group.GID != "" {
-		cmd := types.Command{
-			Exec:     "dscl",
-			Args:     []string{".", "-create", record, "PrimaryGroupID", group.GID},
-			Elevated: true,
-		}
-		if err := system.RunCommand(cmd, initConfig.Variables.Flags.Debug); err != nil {
-			return fmt.Errorf("error modifying group: %v", err)
-		}
-	}
-
-	// Renaming last: every earlier edit addresses the group by its old record name.
-	if group.NewName != "" {
-		cmd := types.Command{
-			Exec:     "dscl",
-			Args:     []string{".", "-change", record, "RecordName", group.Name, group.NewName},
-			Elevated: true,
-		}
-		if err := system.RunCommand(cmd, initConfig.Variables.Flags.Debug); err != nil {
-			return fmt.Errorf("error renaming group: %v", err)
-		}
 	}
 	return nil
 }
@@ -439,207 +312,6 @@ func createUser(user types.User, initConfig *types.InitConfig) error {
 	return nil
 }
 
-func createUserLinux(user types.User, initConfig *types.InitConfig) error {
-	if userExists(user.Name, initConfig, false) {
-		log.Infof("User %s already exists, converging declared attributes instead of creating", user.Name)
-		return convergeUserLinux(user, initConfig)
-	}
-
-	args := []string{"--create-home"}
-	if user.UID != "" {
-		args = append(args, "--uid", user.UID)
-	}
-	if user.Shell != "" {
-		args = append(args, "--shell", user.Shell)
-	}
-	if user.Home != "" {
-		args = append(args, "--home-dir", user.Home)
-	}
-	if user.Comment != "" {
-		args = append(args, "--comment", user.Comment)
-	}
-	if user.System {
-		args = append(args, "--system")
-	}
-	if user.Expire != "" {
-		args = append(args, "--expiredate", user.Expire)
-	}
-	for _, group := range user.Groups {
-		args = append(args, "--groups", group)
-	}
-	args = append(args, user.Name)
-
-	createUserCmd := types.Command{
-		Exec:        "useradd",
-		Args:        args,
-		Elevated:    true,
-		Interactive: helpers.ResolveInteractive(user.Interactive, initConfig.Variables.Flags.Interactive),
-	}
-	if err := system.RunCommand(createUserCmd, initConfig.Variables.Flags.Debug); err != nil {
-		return fmt.Errorf("error creating user: %v", err)
-	}
-	return setLinuxPassword(user, initConfig)
-}
-
-// convergeUserLinux applies a "create" declaration to an account that already
-// exists. usermod is used rather than the NewX fields of modifyUser because the
-// declared home is the intended home, not a request to relocate an existing one.
-func convergeUserLinux(user types.User, initConfig *types.InitConfig) error {
-	args := []string{}
-	if user.UID != "" {
-		args = append(args, "--uid", user.UID)
-	}
-	if user.Shell != "" {
-		args = append(args, "--shell", user.Shell)
-	}
-	if user.Home != "" {
-		args = append(args, "--home", user.Home)
-	}
-	if user.Comment != "" {
-		args = append(args, "--comment", user.Comment)
-	}
-	if user.Expire != "" {
-		args = append(args, "--expiredate", user.Expire)
-	}
-	for _, group := range user.Groups {
-		args = append(args, "--append", "--groups", group)
-	}
-	if len(args) == 0 {
-		log.Debugf("User %s already exists and declares no attributes to converge", user.Name)
-		return setLinuxPassword(user, initConfig)
-	}
-	args = append(args, user.Name)
-
-	cmd := types.Command{
-		Exec:        "usermod",
-		Args:        args,
-		Elevated:    true,
-		Interactive: helpers.ResolveInteractive(user.Interactive, initConfig.Variables.Flags.Interactive),
-	}
-	if err := system.RunCommand(cmd, initConfig.Variables.Flags.Debug); err != nil {
-		return fmt.Errorf("error converging existing user: %v", err)
-	}
-	return setLinuxPassword(user, initConfig)
-}
-
-func createUserDarwin(user types.User, initConfig *types.InitConfig) error {
-	warnUnsupportedMacFields(user)
-
-	if userExists(user.Name, initConfig, false) {
-		log.Infof("User %s already exists, converging declared attributes instead of creating", user.Name)
-		return convergeUserDarwin(user, initConfig)
-	}
-
-	uid := user.UID
-	if uid == "" {
-		allocated, err := macNextFreeID("/Users", "UniqueID", initConfig)
-		if err != nil {
-			return fmt.Errorf("error allocating UID for user %s: %w", user.Name, err)
-		}
-		uid = allocated
-	}
-
-	home := user.Home
-	if home == "" {
-		home = "/Users/" + user.Name
-	}
-	shell := user.Shell
-	if shell == "" {
-		shell = "/bin/zsh"
-	}
-	realName := user.Comment
-	if realName == "" {
-		realName = user.Name
-	}
-
-	interactive := helpers.ResolveInteractive(user.Interactive, initConfig.Variables.Flags.Interactive)
-
-	if commandExists("sysadminctl") {
-		cmd := types.Command{
-			Exec: "sysadminctl",
-			Args: []string{
-				"-addUser", user.Name,
-				"-fullName", realName,
-				"-UID", uid,
-				"-shell", shell,
-				"-home", home,
-			},
-			Elevated:    true,
-			Interactive: interactive,
-		}
-		if err := system.RunCommand(cmd, initConfig.Variables.Flags.Debug); err != nil {
-			return fmt.Errorf("error creating user: %v", err)
-		}
-	} else {
-		record := "/Users/" + user.Name
-		steps := [][]string{
-			{".", "-create", record},
-			{".", "-create", record, "UserShell", shell},
-			{".", "-create", record, "RealName", realName},
-			{".", "-create", record, "UniqueID", uid},
-			{".", "-create", record, "PrimaryGroupID", macDefaultPrimaryGroupID},
-			{".", "-create", record, "NFSHomeDirectory", home},
-		}
-		for _, args := range steps {
-			cmd := types.Command{Exec: "dscl", Args: args, Elevated: true, Interactive: interactive}
-			if err := system.RunCommand(cmd, initConfig.Variables.Flags.Debug); err != nil {
-				return fmt.Errorf("error creating user: %v", err)
-			}
-		}
-	}
-
-	// sysadminctl populates the home directory itself only in some releases, and
-	// the dscl path never does; createhomedir is idempotent, so run it either way.
-	createHomeCmd := types.Command{
-		Exec:     "createhomedir",
-		Args:     []string{"-c", "-u", user.Name},
-		Elevated: true,
-	}
-	if err := system.RunCommand(createHomeCmd, initConfig.Variables.Flags.Debug); err != nil {
-		return fmt.Errorf("error creating home directory for user %s: %v", user.Name, err)
-	}
-
-	if err := macSetPassword(user, initConfig); err != nil {
-		return err
-	}
-
-	return macAddGroups(user.Name, user.Groups, initConfig)
-}
-
-func convergeUserDarwin(user types.User, initConfig *types.InitConfig) error {
-	record := "/Users/" + user.Name
-	attrs := [][2]string{}
-	if user.Shell != "" {
-		attrs = append(attrs, [2]string{"UserShell", user.Shell})
-	}
-	if user.Comment != "" {
-		attrs = append(attrs, [2]string{"RealName", user.Comment})
-	}
-	if user.UID != "" {
-		attrs = append(attrs, [2]string{"UniqueID", user.UID})
-	}
-	if user.Home != "" {
-		attrs = append(attrs, [2]string{"NFSHomeDirectory", user.Home})
-	}
-
-	for _, attr := range attrs {
-		cmd := types.Command{
-			Exec:     "dscl",
-			Args:     []string{".", "-create", record, attr[0], attr[1]},
-			Elevated: true,
-		}
-		if err := system.RunCommand(cmd, initConfig.Variables.Flags.Debug); err != nil {
-			return fmt.Errorf("error converging existing user: %v", err)
-		}
-	}
-
-	if err := macSetPassword(user, initConfig); err != nil {
-		return err
-	}
-
-	return macAddGroups(user.Name, user.Groups, initConfig)
-}
-
 func modifyUser(user types.User, initConfig *types.InitConfig) error {
 	switch userGOOS {
 	case "linux":
@@ -650,152 +322,6 @@ func modifyUser(user types.User, initConfig *types.InitConfig) error {
 		log.Warnf("Modifying users is not supported on Windows")
 	default:
 		return fmt.Errorf("unsupported operating system: %s", userGOOS)
-	}
-	return nil
-}
-
-func modifyUserLinux(user types.User, initConfig *types.InitConfig) error {
-	modifyUserCmd := types.Command{
-		Exec:        "usermod",
-		Args:        []string{user.Name},
-		Elevated:    true,
-		Interactive: helpers.ResolveInteractive(user.Interactive, initConfig.Variables.Flags.Interactive),
-	}
-	if user.NewName != "" {
-		modifyUserCmd.Args = append(modifyUserCmd.Args, "--login", user.NewName)
-	}
-	if user.NewHome != "" {
-		modifyUserCmd.Args = append(modifyUserCmd.Args, "--move-home", "--home", user.NewHome)
-	}
-	if user.NewShell != "" {
-		modifyUserCmd.Args = append(modifyUserCmd.Args, "--shell", user.NewShell)
-	}
-	if user.Comment != "" {
-		modifyUserCmd.Args = append(modifyUserCmd.Args, "--comment", user.Comment)
-	}
-	if user.UID != "" {
-		modifyUserCmd.Args = append(modifyUserCmd.Args, "--uid", user.UID)
-	}
-	if user.Expire != "" {
-		modifyUserCmd.Args = append(modifyUserCmd.Args, "--expiredate", user.Expire)
-	}
-	if user.Lock {
-		modifyUserCmd.Args = append(modifyUserCmd.Args, "--lock")
-	}
-	if user.Unlock {
-		modifyUserCmd.Args = append(modifyUserCmd.Args, "--unlock")
-	}
-	if len(user.AddGroups) > 0 {
-		for _, group := range user.AddGroups {
-			modifyUserCmd.Args = append(modifyUserCmd.Args, "--append", "--groups", group)
-		}
-	}
-
-	err := system.RunCommand(modifyUserCmd, initConfig.Variables.Flags.Debug)
-	if err != nil {
-		return fmt.Errorf("error modifying user: %v", err)
-	}
-
-	if err := setLinuxPassword(user, initConfig); err != nil {
-		return err
-	}
-
-	// Remove groups via gpasswd (usermod doesn't support removing individual groups)
-	for _, group := range user.RemoveGroups {
-		removeGroupCmd := types.Command{
-			Exec:     "gpasswd",
-			Args:     []string{"--delete", user.Name, group},
-			Elevated: true,
-		}
-		if err := system.RunCommand(removeGroupCmd, initConfig.Variables.Flags.Debug); err != nil {
-			return fmt.Errorf("error removing user %s from group %s: %v", user.Name, group, err)
-		}
-	}
-	return nil
-}
-
-func modifyUserDarwin(user types.User, initConfig *types.InitConfig) error {
-	warnUnsupportedMacFields(user)
-
-	record := "/Users/" + user.Name
-	attrs := [][2]string{}
-	if user.NewShell != "" {
-		attrs = append(attrs, [2]string{"UserShell", user.NewShell})
-	} else if user.Shell != "" {
-		attrs = append(attrs, [2]string{"UserShell", user.Shell})
-	}
-	if user.Comment != "" {
-		attrs = append(attrs, [2]string{"RealName", user.Comment})
-	}
-	if user.UID != "" {
-		attrs = append(attrs, [2]string{"UniqueID", user.UID})
-	}
-	if user.NewHome != "" {
-		attrs = append(attrs, [2]string{"NFSHomeDirectory", user.NewHome})
-	} else if user.Home != "" {
-		attrs = append(attrs, [2]string{"NFSHomeDirectory", user.Home})
-	}
-
-	interactive := helpers.ResolveInteractive(user.Interactive, initConfig.Variables.Flags.Interactive)
-
-	for _, attr := range attrs {
-		cmd := types.Command{
-			Exec:        "dscl",
-			Args:        []string{".", "-create", record, attr[0], attr[1]},
-			Elevated:    true,
-			Interactive: interactive,
-		}
-		if err := system.RunCommand(cmd, initConfig.Variables.Flags.Debug); err != nil {
-			return fmt.Errorf("error modifying user: %v", err)
-		}
-	}
-
-	if user.NewHome != "" {
-		// dscl only rewrites the record; the directory itself still has to exist.
-		log.Warnf("User %s: macOS does not move an existing home directory; NFSHomeDirectory now points at %s but its contents were not relocated", user.Name, user.NewHome)
-	}
-
-	if err := macSetPassword(user, initConfig); err != nil {
-		return err
-	}
-
-	if user.Lock {
-		cmd := types.Command{Exec: "pwpolicy", Args: []string{"-u", user.Name, "-disableuser"}, Elevated: true}
-		if err := system.RunCommand(cmd, initConfig.Variables.Flags.Debug); err != nil {
-			return fmt.Errorf("error locking user: %v", err)
-		}
-	}
-	if user.Unlock {
-		cmd := types.Command{Exec: "pwpolicy", Args: []string{"-u", user.Name, "-enableuser"}, Elevated: true}
-		if err := system.RunCommand(cmd, initConfig.Variables.Flags.Debug); err != nil {
-			return fmt.Errorf("error unlocking user: %v", err)
-		}
-	}
-
-	if err := macAddGroups(user.Name, user.AddGroups, initConfig); err != nil {
-		return err
-	}
-	for _, group := range user.RemoveGroups {
-		cmd := types.Command{
-			Exec:     "dseditgroup",
-			Args:     []string{"-o", "edit", "-d", user.Name, "-t", "user", group},
-			Elevated: true,
-		}
-		if err := system.RunCommand(cmd, initConfig.Variables.Flags.Debug); err != nil {
-			return fmt.Errorf("error removing user %s from group %s: %v", user.Name, group, err)
-		}
-	}
-
-	// Renaming last: every edit above addresses the account by its old record name.
-	if user.NewName != "" {
-		cmd := types.Command{
-			Exec:     "dscl",
-			Args:     []string{".", "-change", record, "RecordName", user.Name, user.NewName},
-			Elevated: true,
-		}
-		if err := system.RunCommand(cmd, initConfig.Variables.Flags.Debug); err != nil {
-			return fmt.Errorf("error renaming user: %v", err)
-		}
 	}
 	return nil
 }
@@ -814,161 +340,6 @@ func removeUser(user types.User, initConfig *types.InitConfig) error {
 	return nil
 }
 
-func removeUserLinux(user types.User, initConfig *types.InitConfig) error {
-	if !userExists(user.Name, initConfig, true) {
-		log.Infof("User %s does not exist, nothing to remove", user.Name)
-		return nil
-	}
-
-	removeUserCmd := types.Command{
-		Exec:        "userdel",
-		Args:        []string{user.Name},
-		Elevated:    true,
-		Interactive: helpers.ResolveInteractive(user.Interactive, initConfig.Variables.Flags.Interactive),
-	}
-	if user.RemoveHome {
-		removeUserCmd.Args = append(removeUserCmd.Args, "--remove")
-	}
-
-	if err := system.RunCommand(removeUserCmd, initConfig.Variables.Flags.Debug); err != nil {
-		return fmt.Errorf("error removing user: %v", err)
-	}
-	return nil
-}
-
-func removeUserDarwin(user types.User, initConfig *types.InitConfig) error {
-	if !userExists(user.Name, initConfig, true) {
-		log.Infof("User %s does not exist, nothing to remove", user.Name)
-		return nil
-	}
-
-	interactive := helpers.ResolveInteractive(user.Interactive, initConfig.Variables.Flags.Interactive)
-
-	if commandExists("sysadminctl") {
-		// sysadminctl deletes the home directory unless told otherwise, which is
-		// the opposite of userdel's default; -keepHome restores parity with the
-		// blueprint's remove_home flag.
-		args := []string{"-deleteUser", user.Name}
-		if !user.RemoveHome {
-			args = append(args, "-keepHome")
-		}
-		cmd := types.Command{Exec: "sysadminctl", Args: args, Elevated: true, Interactive: interactive}
-		if err := system.RunCommand(cmd, initConfig.Variables.Flags.Debug); err != nil {
-			return fmt.Errorf("error removing user: %v", err)
-		}
-		return nil
-	}
-
-	cmd := types.Command{
-		Exec:        "dscl",
-		Args:        []string{".", "-delete", "/Users/" + user.Name},
-		Elevated:    true,
-		Interactive: interactive,
-	}
-	if err := system.RunCommand(cmd, initConfig.Variables.Flags.Debug); err != nil {
-		return fmt.Errorf("error removing user: %v", err)
-	}
-	if user.RemoveHome {
-		log.Warnf("User %s: remove_home requires sysadminctl, which is not available; the home directory was left in place", user.Name)
-	}
-	return nil
-}
-
-// ---------------------------------------------------------------------------
-// macOS helpers
-// ---------------------------------------------------------------------------
-
-// warnUnsupportedMacFields reports the blueprint fields that have no Open
-// Directory equivalent, rather than dropping them silently and letting the run
-// look like it applied them.
-func warnUnsupportedMacFields(user types.User) {
-	if user.System {
-		log.Warnf("User %s: 'system' is ignored on macOS; UIDs below %d are reserved by Apple and are not allocated by rwr", user.Name, macMinID)
-	}
-	if user.Expire != "" {
-		log.Warnf("User %s: 'expire' is ignored on macOS; local Open Directory accounts have no expiration date field", user.Name)
-	}
-}
-
-func macAddGroups(name string, groups []string, initConfig *types.InitConfig) error {
-	for _, group := range groups {
-		if macUserInGroup(name, group, initConfig) {
-			log.Debugf("User %s is already a member of group %s", name, group)
-			continue
-		}
-		cmd := types.Command{
-			Exec:     "dseditgroup",
-			Args:     []string{"-o", "edit", "-a", name, "-t", "user", group},
-			Elevated: true,
-		}
-		if err := system.RunCommand(cmd, initConfig.Variables.Flags.Debug); err != nil {
-			return fmt.Errorf("error adding user %s to group %s: %v", name, group, err)
-		}
-	}
-	return nil
-}
-
-// macSetPassword sets the account password through dscl.
-//
-// macOS stores a salted SHA-512 PBKDF2 blob it computes itself, so unlike Linux
-// it needs the cleartext password — there is no hash to pre-compute. That value
-// therefore appears in the argv of a sudo'd process, visible in `ps` to every
-// local user and recorded in sudo's syslog entry.
-//
-// TODO: feed the password to `dscl . -passwd` on stdin instead.
-// types.Command.Stdin and its RunCommand wiring exist now (this file already
-// uses them for the Linux chpasswd path below); what remains is confirming
-// dscl reads a password from stdin non-interactively on current macOS.
-func macSetPassword(user types.User, initConfig *types.InitConfig) error {
-	if user.Password == "" {
-		return nil
-	}
-	log.Warnf("User %s: the password is passed as a command argument on macOS and is briefly visible to other local users in the process list", user.Name)
-
-	cmd := types.Command{
-		Exec:     "dscl",
-		Args:     []string{".", "-passwd", "/Users/" + user.Name, user.Password},
-		Elevated: true,
-		// dscl offers no way to read the password from stdin, so it cannot be kept
-		// out of argv the way chpasswd allows on Linux. It can at least be kept out
-		// of the debug and dry-run log lines, which print the whole argv.
-		Secrets: []string{user.Password},
-	}
-	if err := system.RunCommand(cmd, initConfig.Variables.Flags.Debug); err != nil {
-		// The error is deliberately not wrapped with anything derived from the
-		// password value.
-		return fmt.Errorf("error setting password for user %s: %v", user.Name, err)
-	}
-	return nil
-}
-
-// macNextFreeID picks the lowest unused numeric ID at or above macMinID by
-// reading the IDs Open Directory already hands out.
-func macNextFreeID(path, attribute string, initConfig *types.InitConfig) (string, error) {
-	cmd := types.Command{Exec: "dscl", Args: []string{".", "-list", path, attribute}}
-	out, err := system.RunCommandOutput(cmd, initConfig.Variables.Flags.Debug)
-	if err != nil {
-		return "", fmt.Errorf("error listing %s %s: %w", path, attribute, err)
-	}
-
-	used := map[int]bool{}
-	for _, line := range strings.Split(out, "\n") {
-		fields := strings.Fields(line)
-		if len(fields) < 2 {
-			continue
-		}
-		if id, convErr := strconv.Atoi(fields[len(fields)-1]); convErr == nil {
-			used[id] = true
-		}
-	}
-
-	id := macMinID
-	for used[id] {
-		id++
-	}
-	return strconv.Itoa(id), nil
-}
-
 // ---------------------------------------------------------------------------
 // password handling
 // ---------------------------------------------------------------------------
@@ -979,60 +350,6 @@ var modernCryptHash = regexp.MustCompile(`^\$[A-Za-z0-9_-]{1,12}\$`)
 
 // desCryptHash matches the legacy 13-character DES hash.
 var desCryptHash = regexp.MustCompile(`^[./A-Za-z0-9]{13}$`)
-
-// setLinuxPassword sets the account password by piping "<name>:<value>" to
-// chpasswd on standard input.
-//
-// The obvious approach — useradd/usermod --password — is wrong twice over. That
-// flag writes its argument verbatim into the hash field of /etc/shadow, so a
-// cleartext value does not become the account's password; it becomes a hash
-// nothing can ever match and the account silently has none. And because the
-// value is an argument of a sudo'd process it is readable in `ps` by every local
-// user for the lifetime of the call, and lands in sudo's syslog record.
-//
-// Stdin avoids both: chpasswd hashes cleartext itself, `chpasswd -e` accepts a
-// pre-computed hash, and neither form puts the secret in argv or in the debug
-// log, which prints cmd.Args.
-func setLinuxPassword(user types.User, initConfig *types.InitConfig) error {
-	if user.Password == "" {
-		return nil
-	}
-
-	name := user.Name
-	if user.NewName != "" {
-		// A rename in the same declaration has already been applied.
-		name = user.NewName
-	}
-
-	args := []string{}
-	if isCryptHash(user.Password) {
-		args = append(args, "-e")
-	}
-
-	cmd := types.Command{
-		Exec:     "chpasswd",
-		Args:     args,
-		Elevated: true,
-		Stdin:    name + ":" + user.Password + "\n",
-	}
-	if err := system.RunCommand(cmd, initConfig.Variables.Flags.Debug); err != nil {
-		// The error deliberately excludes the value: it is the secret.
-		return fmt.Errorf("error setting password for user %s: %v", name, err)
-	}
-
-	log.Debugf("Password set for user %s", name)
-	return nil
-}
-
-// isCryptHash reports whether value is already a crypt(3) hash, which decides
-// whether chpasswd is told to hash it or to take it as-is.
-func isCryptHash(value string) bool {
-	switch value {
-	case "!", "!!", "*", "*LK*":
-		return true
-	}
-	return modernCryptHash.MatchString(value) || desCryptHash.MatchString(value)
-}
 
 func processGroupImports(items []types.Group, blueprintDir string, format string, treeVersion int) ([]types.Group, error) {
 	return helpers.ResolveImports(items, blueprintDir,
